tests/conftest.py
"""Test configuration"""
import pytest
from fastapi.testclient import TestClient
from src.main import app

client = TestClient(app)


@pytest.fixture
def test_client():
    """Test client fixture"""
    return client


@pytest.fixture
def test_user_data():
    """Test user data"""
    return {
        "username": "testuser",
        "email": "test@example.com",
        "password": "TestPassword123!",
        "phone": "1234567890"
    }
