requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.8.0
bcrypt==4.1.1
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.0
requests==2.31.0
pyyaml>=6.0

# Local packages
--find-links ../utils-service/dist
utils-service==0.1.1